Refactor genotype pack architecture
